# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/cv": {
      "get": {
        "operationId": "CvController_download",
        "parameters": [
          {
            "name": "persona",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Download the CV PDF (logs the download to CvDownload first)",
        "tags": [
          "cv"
        ]
      }
    },
...
  },
  "info": {
    "title": "Syrine Portfolio API",
    "description": "Backend for the portfolio site (contact, CV, analytics, admin)",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "portfolio_admin"
      }
    },
...
  }
}